+++ src/test/scala/org/mmisw/orr/ont/app/SequenceSpec.scala
@@ -475,6 +475,37 @@ class SequenceSpec extends MutableScalatraSpec with BaseSpec with Mockito with L
     }
   }
 
+  "Upload file with '_guess' format (POST /ont/upload)" should {
+    "succeed for N3 contents and return expected info" in {
+      val file = new File("src/test/resources/core_variable.n3")
+      post("/ont/upload", Map("format" -> "_guess", "visibility" -> "public"),
+        Map("file" -> file), headers = userHeaders
+      ) {
+        status must_== 200
+        val b = body
+        println(s"upload response body=$b")
+        val uploadedFileInfo = parse(b).extract[UploadedFileInfo]
+        println(s"uploadedFileInfo=$uploadedFileInfo")
+        uploadedFileInfo.userName must_== userName
+        uploadedFileInfo.format must_== "n3"
+      }
+    }
+    "succeed for RDF contents and return expected info" in {
+      val file = new File("src/test/resources/ont1.rdf")
+      post("/ont/upload", Map("format" -> "_guess", "visibility" -> "public"),
+        Map("file" -> file), headers = userHeaders
+      ) {
+        status must_== 200
+        val b = body
+        println(s"upload response body=$b")
+        val uploadedFileInfo = parse(b).extract[UploadedFileInfo]
+        println(s"uploadedFileInfo=$uploadedFileInfo")
+        uploadedFileInfo.userName must_== userName
+        uploadedFileInfo.format must_== "rdf"
+      }
+    }
+  }
+
   "Register a new ont (POST /ont) whose file has been previously uploaded" should {
     "succeed with user credentials" in {
       val uploadedOntUri = ont1Uri + "_uploaded"